invalid styles moved to a separate file
