+++ src/shiver/views/invalid_styles.py
@@ -1,0 +1,48 @@
+"""CSS styles invalid fields"""
+INVALID_QLINEEDIT = """
+QLineEdit {
+border-color: red;
+border-style: outset;
+border-width: 2px;
+border-radius: 4px;
+padding-left: -1px;
+padding-right: -1px;
+padding-top: 1px;
+padding-bottom: 1px;
+}
+"""
+INVALID_QTABLEWIDGET = """
+QTableWidget {
+border-color: red;
+border-style: outset;
+border-width: 2px;
+border-radius: 4px;
+padding-left: -1px;
+padding-right: -1px;
+padding-top: 1px;
+padding-bottom: 1px;
+}
+"""
+INVALID_QLISTWIDGET = """
+QListWidget {
+border-color: red;
+border-style: outset;
+border-width: 2px;
+border-radius: 4px;
+padding-left: -1px;
+padding-right: -1px;
+padding-top: 1px;
+padding-bottom: 1px;
+}
+"""
+INVALID_QCHECKBOX = """
+QCheckBox::indicator{
+border-color: red;
+border-style: outset;
+border-width: 2px;
+padding-left: -1px;
+padding-right: -1px;
+padding-top: 1px;
+padding-bottom: 1px;
+}
+"""
