+++ src/shiver/views/sample.py
@@ -23,7 +23,7 @@ except ImportError:
     QString = type("")
 
 from shiver.models.help import help_function
-from .histogram_parameters import INVALID_QLINEEDIT
+from .invalid_styles import INVALID_QLINEEDIT
 
 
 class SampleView(QWidget):
